lec8/requirements.txt
streamlit
python-dotenv
google-generativeai
pdf2image
pillow